+++ mobile/android/android-components/docs/changelog.md
@@ -13,7 +13,7 @@ permalink: /changelog/
 * Compiled against:
   * Android (SDK: 27, Support Libraries: 27.1.1)
   * Kotlin (Stdlib: 1.2.71, Coroutines: 0.30.2)
-  * GeckoView (Nightly: 65.0.20181023100123, Beta: 64.0.20181022150107, Release: 63.0.20181018182531)
+  * GeckoView (Nightly: **65.0.20181107100135** 🔺, Beta: 64.0.20181022150107, Release: 63.0.20181018182531)
 
 * **concept-storage**, **browser-storage-memory**, **browser-storage-sync**
   * Added a `getDomainSuggestion` method to `HistoryStorage` which is intended to power awesomebar-like functionality.